+++ src/v1/ErrorHandling/StructuredProblemReport/StructuredProblemReportData.ts
@@ -32,7 +32,7 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-import type { HttpStatusCode, NodeJSModuleName } from "../..";
+import type { NodeJSModuleName } from "../..";
 import type { NoExtraData, ExtraData } from "../ExtraData";
 
 /**
@@ -75,18 +75,4 @@ export interface StructuredProblemReportData<E extends ExtraData | NoExtraData> 
      * `.extra` contains any unique data for this error occurance.
      */
     readonly extra: E;
-
-    /**
-     * `.status` contains the HTTP status that best fits this error.
-     *
-     * NOTE that this is from the point-of-view of the code that originally
-     * threw this error.
-     *
-     * e.g. a library may report a `422` (validation failure),
-     * but it doesn't know where the rejected input comes from.
-     *
-     * The calling app DOES know, and it may decide to report a `500`
-     * (internal server error) back to the end-user instead.
-     */
-    readonly status: HttpStatusCode;
 }