+++ src/v1/Errors/ObjectHasMissingMethods/ObjectHasMissingMethodsError.ts
@@ -32,7 +32,6 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 import { AppError, AppErrorData, makeStructuredProblemReport } from "../../ErrorHandling";
-import type { HttpStatusCode } from "../../SupportingTypes";
 import { MODULE_NAME } from "../defaults/MODULE_NAME";
 import type { ObjectHasMissingMethodsData } from "./ObjectHasMissingMethodsData";
 
@@ -52,7 +51,6 @@ export class ObjectHasMissingMethodsError extends AppError<ObjectHasMissingMetho
             description: "object has missing methods",
             errorId: params.errorId,
             extra: { public: params.public },
-            status: 500 as HttpStatusCode,
         });
 
         super(spr);