+++ src/v1/Errors/UnsupportedType/UnsupportedTypeError.ts
@@ -32,7 +32,6 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 import { AppError, AppErrorData, makeStructuredProblemReport } from "../../ErrorHandling";
-import type { HttpStatusCode } from "../../SupportingTypes";
 import { MODULE_NAME } from "../defaults/MODULE_NAME";
 import type { UnsupportedTypeData } from "./UnsupportedTypeData";
 
@@ -51,7 +50,6 @@ export class UnsupportedTypeError extends AppError<UnsupportedTypeData> {
             extra: {
                 public: params.public
             },
-            status: 422 as HttpStatusCode,
         });
 
         super(spr);